webapp/requirements.txt
allure-behave==2.9.45
allure-python-commons==2.9.45
attrs==21.4.0
behave==1.2.6
behave-webdriver==0.3.1
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
importlib-metadata==4.11.0
parse==1.19.0
parse-type==0.6.0
pluggy==1.0.0
requests==2.27.1
selenium==3.141.0
six==1.16.0
typing-extensions==4.0.1
urllib3==1.26.8
zipp==3.7.0
webdriver_manager